Properly support scrolling skies in `MAPINFO` lumps

+++ src/p_setup.h
@@ -96,6 +96,6 @@ int P_GetMapPar(const int ep, const int map);
 bool P_GetMapPistolStart(const int ep, const int map);
 int P_GetMapSecretNext(const int ep, const int map);
 int P_GetMapSky1Texture(const int ep, const int map);
-int P_GetMapSky1ScrollDelta(const int ep, const int map);
+float P_GetMapSky1ScrollDelta(const int ep, const int map);
 int P_GetMapTitlePatch(const int ep, const int map);
 int P_GetAllowMonsterTelefrags(const int ep, const int map);
